Extract add_to_transmission
